Allow the user to avoid stack URL redirects
If you go to the path ../stacks/<stack_id>, you are (supposed to be)
redirected to the proper path, ../stacks/<stack_name>/<stack_id>.
Whether it's a bug with our deployment or OpenStack in general,
the redirect for stack updates is broken. This allows the user to
go directly to the correct path.